requirements.txt
streamlit>=1.35.0
snowflake-connector-python[pandas]>=3.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
PyYAML>=6.0.1
